src/test/javascript/spec/app/entities/client-account/client-account.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { ClientAccountService } from 'app/entities/client-account/client-account.service';
import { IClientAccount, ClientAccount } from 'app/shared/model/client-account.model';

describe('Service Tests', () => {
  describe('ClientAccount Service', () => {
    let injector: TestBed;
    let service: ClientAccountService;
    let httpMock: HttpTestingController;
    let elemDefault: IClientAccount;
    let expectedResult: IClientAccount | IClientAccount[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(ClientAccountService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new ClientAccount(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', false, 'image/png', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a ClientAccount', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new ClientAccount()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a ClientAccount', () => {
        const returnedFromService = Object.assign(
          {
            phone: 'BBBBBB',
            referal: 'BBBBBB',
            referedBy: 'BBBBBB',
            miniBio: 'BBBBBB',
            verifiedPhone: true,
            photo: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of ClientAccount', () => {
        const returnedFromService = Object.assign(
          {
            phone: 'BBBBBB',
            referal: 'BBBBBB',
            referedBy: 'BBBBBB',
            miniBio: 'BBBBBB',
            verifiedPhone: true,
            photo: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a ClientAccount', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
